Aggregate EqAsset logic to class
