+++ frontend/src/app/eq-assets/emitters.json
