+++ frontend/src/app/eq-assets/item-icons-map.json
