+++ frontend/src/app/eq-assets/monograms-map.json
