+++ frontend/src/app/eq-assets/npc-models-map.json
