+++ frontend/src/app/eq-assets/objects-map.json
