+++ frontend/src/app/eq-assets/player-animations.json
