+++ frontend/src/app/eq-assets/spell-animations-map.json
