+++ frontend/src/app/eq-assets/spell-icons-map.json
